+++ crates/m80-net-outbound/src/iptables.rs
@@ -352,7 +352,7 @@ fn ensure_forwarding_entry_rules(
         "FORWARD",
         vec![
             "-i".into(),
-            state.tap_name.clone(),
+            state.bridge.bridge_name.clone(),
             "-s".into(),
             guest.clone(),
             "-m".into(),
@@ -370,7 +370,7 @@ fn ensure_forwarding_entry_rules(
         "FORWARD",
         vec![
             "-o".into(),
-            state.tap_name.clone(),
+            state.bridge.bridge_name.clone(),
             "-d".into(),
             guest.clone(),
             "-m".into(),
@@ -388,7 +388,7 @@ fn ensure_forwarding_entry_rules(
         "FORWARD",
         vec![
             "-o".into(),
-            state.tap_name.clone(),
+            state.bridge.bridge_name.clone(),
             "-d".into(),
             guest,
             "-m".into(),